Conflicts:
	Erebus/Erebus.vcxproj
	Erebus/Erebus.vcxproj.filters
	Erebus/Scripts/fireball.lua
	Erebus/Scripts/player.lua
	Erebus/Scripts/projectileType.lua
	Erebus/Scripts/spellUtility.lua
	Gear/RenderQueue.cpp

+++ Nurn/DamagePacket.hpp
@@ -7,7 +7,7 @@ union DamagePacket
 	struct DamagePacketStruct
 	{
 		uint16_t transformID;
-		uint16_t damage;
+		float damage;
 
 	}data;
 
@@ -15,7 +15,7 @@ union DamagePacket
 	{
 	}
 
-	DamagePacket(const uint16_t &transformID, const uint16_t &damage)
+	DamagePacket(const uint16_t &transformID, const float &damage)
 	{
 		data.transformID = transformID;
 		data.damage = damage;
